소스 보기 기능 추가

+++ app/src/main/java/com/aiden/andmodule/activity/WebActivity.java
@@ -2,12 +2,16 @@ package com.aiden.andmodule.activity;
 
 import android.app.Dialog;
 import android.content.DialogInterface;
+import android.content.Intent;
 import android.graphics.Bitmap;
+import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Message;
 import android.util.Log;
 import android.view.KeyEvent;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.webkit.JsResult;
 import android.webkit.WebChromeClient;
@@ -32,6 +36,7 @@ public class WebActivity extends AppCompatActivity {
     String TAG = "WebActivity";
     private WebView webView;
     private String URL = "https://m.edu2080.co.kr/review?isapp=y";
+    String source_url = "https://github.com/leeminyong/simpleapp/blob/master/app/src/main/java/com/aiden/andmodule/activity/WebActivity.java";
 
     SwipeRefreshLayout swipe;
     ProgressBar progressBar;
@@ -54,8 +59,8 @@ public class WebActivity extends AppCompatActivity {
         });
 
         // WebView의 설정
-      webView.setWebChromeClient(new CustomWebChromelient());
-      webView.setWebViewClient(new CustomWebViewClient());
+        webView.setWebChromeClient(new CustomWebChromelient());
+        webView.setWebViewClient(new CustomWebViewClient());
     }
 
     public void LoadWeb() {
@@ -74,9 +79,29 @@ public class WebActivity extends AppCompatActivity {
     }
 
     @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.code_list, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == R.id.action_setting) {
+            LogUtil.e(TAG, "show source");
+            startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(source_url)));
+        }
+        return super.onOptionsItemSelected(item);
+    }
+    @Override
     protected void onPause() {
         super.onPause();
         callHiddenWebViewMethod("onPause");
+
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
         if (webView != null) {
             LogUtil.e(TAG, "it is call Destory of WebView ");
             webView.destroy();
@@ -84,25 +109,26 @@ public class WebActivity extends AppCompatActivity {
         }
     }
 
-
-
     //key down
     @Override
     public boolean onKeyDown(int keyCode, KeyEvent event) {
-        if ((keyCode == KeyEvent.KEYCODE_BACK) && webView.canGoBack()) {
-            Toast.makeText(this, "뒤로가기  이벤트", Toast.LENGTH_SHORT).show();
-            webView.goBack();
-            return true;
-        }
-        //백할 페이가 없다면
-        if ((keyCode == KeyEvent.KEYCODE_BACK) && (webView.canGoBack() == false)) {
-            Toast.makeText(this, "버튼 클릭 이벤트", Toast.LENGTH_SHORT).show();
-            //다이아로그박스 출력
-            new AlertDialog.Builder(this)
-                    .setTitle("프로그램 종료")
-                    .setMessage("프로그램을 종료하시겠습니까?")
-                    .setPositiveButton("예", (dialog, which) -> android.os.Process.killProcess(android.os.Process.myPid()))
-                    .setNegativeButton("아니오", null).show();
+
+        if (webView != null) {
+            if ((keyCode == KeyEvent.KEYCODE_BACK) && webView.canGoBack()) {
+                Toast.makeText(this, "뒤로가기  이벤트", Toast.LENGTH_SHORT).show();
+                webView.goBack();
+                return true;
+            }
+            //백할 페이가 없다면
+            if ((keyCode == KeyEvent.KEYCODE_BACK) && (webView.canGoBack() == false)) {
+                Toast.makeText(this, "버튼 클릭 이벤트", Toast.LENGTH_SHORT).show();
+                //다이아로그박스 출력
+                new AlertDialog.Builder(this)
+                        .setTitle("프로그램 종료")
+                        .setMessage("프로그램을 종료하시겠습니까?")
+                        .setPositiveButton("예", (dialog, which) -> android.os.Process.killProcess(android.os.Process.myPid()))
+                        .setNegativeButton("아니오", null).show();
+            }
         }
         return super.onKeyDown(keyCode, event);
     }
@@ -113,6 +139,7 @@ public class WebActivity extends AppCompatActivity {
         super.onResume();
         callHiddenWebViewMethod("onResume");
     }
+
     private void callHiddenWebViewMethod(String name) {
         if (webView != null) {
             try {
@@ -129,6 +156,7 @@ public class WebActivity extends AppCompatActivity {
             }
         }
     }
+
     protected class CustomWebChromelient extends WebChromeClient {
         @Override
         public boolean onJsAlert(WebView view, String url, String message, final JsResult result) {
@@ -136,7 +164,7 @@ public class WebActivity extends AppCompatActivity {
                     .setMessage(message)
                     .setPositiveButton(android.R.string.ok,
                             (dialog, which) -> result.confirm()).setNegativeButton(android.R.string.cancel,
-                            (dialog, which) -> result.cancel()).setCancelable(false).create().show();
+                    (dialog, which) -> result.cancel()).setCancelable(false).create().show();
             return true;
         }
 
@@ -147,15 +175,15 @@ public class WebActivity extends AppCompatActivity {
                     .setMessage(message)
                     .setPositiveButton(android.R.string.ok,
                             (dialog, which) -> result.confirm()).setNegativeButton(android.R.string.cancel,
-                            (dialog, which) -> result.cancel()).setCancelable(false).create().show();
+                    (dialog, which) -> result.cancel()).setCancelable(false).create().show();
 
-                    return true;
+            return true;
         }
 
-  String userAgent;
+        String userAgent;
 
-   @Override
-   public boolean onCreateWindow(WebView view, boolean isDialog, boolean isUserGesture, Message resultMsg) {
+        @Override
+        public boolean onCreateWindow(WebView view, boolean isDialog, boolean isUserGesture, Message resultMsg) {
             WebView newWebView = new WebView(WebActivity.this);
             WebSettings webSettings = newWebView.getSettings();
             webSettings.setJavaScriptEnabled(true);
@@ -168,10 +196,10 @@ public class WebActivity extends AppCompatActivity {
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
                 newWebView.getSettings().setMixedContentMode(WebSettings.MIXED_CONTENT_ALWAYS_ALLOW);
             }
-            newWebView.setWebChromeClient(new WebChromeClient(){
+            newWebView.setWebChromeClient(new WebChromeClient() {
                 @Override
                 public void onCloseWindow(WebView window) {
-                    Log.e(TAG,"------------------------------1");
+                    Log.e(TAG, "------------------------------1");
                     dialog.dismiss();
                 }
             });
@@ -180,6 +208,7 @@ public class WebActivity extends AppCompatActivity {
                 public boolean shouldOverrideUrlLoading(WebView view, String url) {
                     return false;
                 }
+
                 @Override
                 public void onLoadResource(WebView view, String url) {
                     super.onLoadResource(view, url);
@@ -196,18 +225,22 @@ public class WebActivity extends AppCompatActivity {
         @Override
         public boolean shouldOverrideUrlLoading(WebView view, String url) {
             LogUtil.e(TAG, "(WebViewClient)should ---->" + url);
-            return  false;
+            return false;
         }
+
         @Override
         public void onPageStarted(WebView view, String url, Bitmap favicon) {
             super.onPageStarted(view, url, favicon);
             progressBar.setVisibility(View.VISIBLE);
         }
+
         public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
             webView.loadUrl("file:///android_asset/error.html");
         }
+
         public void onLoadResource(WebView view, String url) {
         }
+
         public void onPageFinished(WebView view, String url) {
             //Hide the SwipeReefreshLayout
             progressBar.setVisibility(View.GONE);
